+++ auth-utils/src/authentication/index.ts
@@ -1,3 +1,0 @@
-export { default } from './authentication';
-
-export * from './authentication';